+++ src/main/java/net/litetex/authback/mixin/common/ServicesMixin.java
@@ -10,13 +10,16 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import com.mojang.authlib.GameProfileRepository;
 import com.mojang.authlib.minecraft.SessionService;
 import com.mojang.authlib.services.MinecraftServicesDiscoveryService;
+import com.mojang.authlib.services.ServicesKeySet;
 
 import net.litetex.authback.common.AuthBackCommon;
+import net.litetex.authback.common.access.MinecraftServicesSessionServiceExt;
 import net.minecraft.server.Services;
 import net.minecraft.server.players.ProfileResolver;
 import net.minecraft.server.players.UserNameToIdResolver;
 
 
+@SuppressWarnings("checkstyle:IllegalIdentifierName")
 @Mixin(Services.class)
 public abstract class ServicesMixin
 {
@@ -32,7 +35,14 @@ public abstract class ServicesMixin
 		final File nameCacheDir,
 		final CallbackInfoReturnable<Services> cir)
 	{
-		final SessionService minecraftSessionService = serviceAccess.createMinecraftSessionService();
+		// As of 26.3 this is no longer a getter but creates/queries the backend!
+		final ServicesKeySet servicesKeySet = serviceAccess.getServicesKeySet();
+		
+		// 26.3
+		// Do not use createMinecraftSessionService as this calls getServicesKeySet again
+		// -> The fetch is executed twice
+		final SessionService minecraftSessionService =
+			new MinecraftServicesSessionServiceExt(servicesKeySet, serviceAccess.getProxy(), serviceAccess);
 		final GameProfileRepository gameProfileRepository = serviceAccess.createProfileRepository();
 		final UserNameToIdResolver userNameToIdResolver =
 			AuthBackCommon.instance().createUserNameToIdResolver(gameProfileRepository, nameCacheDir);
@@ -41,7 +51,7 @@ public abstract class ServicesMixin
 			userNameToIdResolver);
 		cir.setReturnValue(new Services(
 			minecraftSessionService,
-			serviceAccess.getServicesKeySet(),
+			servicesKeySet,
 			gameProfileRepository,
 			userNameToIdResolver,
 			profileResolver));